core --> bin rename folder
