+++ bin/basic_training.py
